requirements.txt
aiofiles==24.1.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
asyncpg==0.30.0
bcrypt==4.0.1
billiard==4.2.2
black==25.1.0
celery==5.5.3
cffi==2.0.0
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dill==0.4.0
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
filelock==3.19.1
flower==2.0.1
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
humanize==4.13.0
identify==2.6.14
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
kombu==5.5.4
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pre_commit==4.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
ruff==0.12.12
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tomlkit==0.13.3
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
zope.event==6.0
zope.interface==8.0
